[#845] Organize implementation into structs
