+++ iceoryx2-tunnels/core/src/tunnel/discovery/mod.rs
@@ -10,8 +10,5 @@
 //
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-mod discovery;
-mod ports;
-mod tunnel;
-
-pub use tunnel::*;
+pub(crate) mod subscriber;
+pub(crate) mod tracker;
