+++ iceoryx2-tunnels/core/src/tunnel/ports/event.rs
@@ -10,8 +10,12 @@
 //
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-mod discovery;
-mod ports;
-mod tunnel;
+use iceoryx2::{
+    port::{listener::Listener, notifier::Notifier},
+    service::Service,
+};
 
-pub use tunnel::*;
+pub(crate) struct Ports<S: Service> {
+    pub(crate) notifier: Notifier<S>,
+    pub(crate) listener: Listener<S>,
+}
